Bootcamp project_Vinay Araveeti/HealthBot_Streamlit/requirements.txt

streamlit==1.55.0
langchain==0.2.16
langchain_openai==0.1.23
langgraph==0.2.19
langchainhub==0.1.21
tavily-python==0.4.0
langchain-community==0.2.16
python-dotenv==1.0.1
cohere>=5.0.0
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0